Dynamically discover I2C bus numbers
The previous scheme statically assigned bus numbers which assumed that
the primary I2C bus is 0 and the virtual buses created by the I2C switch
would start from 1.  On the wp76xx, the primary i2c bus is 4, so this
didn't work.  The new version of the code still assumes that the virtual
buses follow the primary bus, but now the primary bus number is
discoverd.  This change also eliminates the pin spec type and introduces
an expander identifier type.  It seems to be more convenient to identify
the pin and expander separately.